Feature/flush log cache (#10)
This change introduces a new custom setting called Flush Log Cache Level. It works in association of the reporting log level. If the log event is published and the level of said event is equals or higher to the configured Flush Log Cache Level then the log cache will be cleared.

NOTE: This will NOT effect the Batched Logger because it relies on the cache for the current batching logic.